Feature: Configured JSON Schema generation
