+++ contract/Cargo.toml
@@ -20,6 +20,7 @@ overflow-checks = true
 
 [features]
 default = []
+library = []
 # for quicker tests, cargo test --lib
 # for more explicit tests, cargo test --features=backtraces
 backtraces = ["cosmwasm-std/backtraces"]
@@ -27,10 +28,10 @@ backtraces = ["cosmwasm-std/backtraces"]
 [dependencies]
 cosmwasm-std = { package = "secret-cosmwasm-std", version = "1.1" }
 cosmwasm-storage = { package = "secret-cosmwasm-storage", version = "1.1" }
-schemars = { version = "0.8" }
+schemars = { version = "1.0" }
 serde = { version = "1.0" }
 thiserror = { version = "2.0" }
-cosmwasm-schema = "2.2"
+cosmwasm-schema = "3.0"
 secret-toolkit = { version = "0.10", features = ["permit"] }
 cw-storage-plus = { version = "2.0", default-features = false }
 secret-toolkit-storage = "0.10"
@@ -42,4 +43,4 @@ default-features = false
 
 [[bin]]
 name = "schema"
-required-features = ["schema"]
+# required-features = ["schema"]
