+++ contract/src/bin/schema.rs
@@ -1,9 +1,9 @@
 use std::env::current_dir;
 use std::fs::create_dir_all;
 
-use cosmwasm_schema::{export_schema, remove_schemas, schema_for};
+use cosmwasm_schema::{export_schema, remove_schemas, schema_for, write_api};
 
-use secret_contract_example::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
+use chain_poker::msg::{ExecuteMsg, InstantiateMsg, QueryMsg};
 
 fn main() {
     let mut out_dir = current_dir().unwrap();
