+++ contract/src/contract/query.rs
@@ -1,10 +1,11 @@
 use cosmwasm_std::{to_binary, Binary, CanonicalAddr, Deps, Env, StdError, StdResult};
 use secret_toolkit::permit::Permit;
 
+use crate::msg::{InGameStatus, LobbyStatus};
 use crate::state::{
-    game_state::GameState, get_balances, InGameStatus, LobbyStatus, ADMIN, ALL_PLAYERS, BETS,
-    BUTTON_POSITION, CURRENT_MIN_BET, CURRENT_STATE, CURRENT_TURN_POSITION, HANDS, LOBBY_CONFIG,
-    POT, TABLE, USERNAMES,
+    game_state::GameState, get_balances, ADMIN, ALL_PLAYERS, BETS, BUTTON_POSITION,
+    CURRENT_MIN_BET, CURRENT_STATE, CURRENT_TURN_POSITION, HANDS, LOBBY_CONFIG, POT, TABLE,
+    USERNAMES,
 };
 
 pub fn query_pre_start_state(deps: Deps) -> StdResult<Binary> {
