+++ contract/src/msg.rs
@@ -1,9 +1,10 @@
+use cosmwasm_schema::{cw_serde, QueryResponses};
 use cosmwasm_std::Uint128;
-use schemars::JsonSchema;
 use secret_toolkit::permit::Permit;
-use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, JsonSchema)]
+use crate::state::LobbyConfig;
+
+#[cw_serde]
 pub struct InstantiateMsg {
     pub username: String,
     pub big_blind: u32,
@@ -11,8 +12,7 @@ pub struct InstantiateMsg {
     pub min_buy_in_bb: u8,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
 pub enum ExecuteMsg {
     StartGame {},
     BuyIn { username: String },
@@ -23,9 +23,31 @@ pub enum ExecuteMsg {
     Withdraw {},
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug, Eq, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
+#[cw_serde]
+#[derive(QueryResponses)]
 pub enum QueryMsg {
-    ViewPreStartState {},
-    ViewGameState { permit: Permit },
+    #[returns(LobbyStatus)]
+    ViewLobbyStatus {},
+
+    #[returns(InGameStatus)]
+    ViewGameStatus { permit: Permit },
+}
+
+#[cw_serde]
+pub struct LobbyStatus {
+    pub admin: String,
+    pub lobby_config: LobbyConfig,
+    pub is_started: bool,
+    pub balances: Vec<(String, u128)>,
+}
+
+#[cw_serde]
+pub struct InGameStatus {
+    pub balances: Vec<(String, u128)>,
+    pub table: Vec<u8>,
+    pub pot: u128,
+    pub hand: Option<(u8, u8)>,
+    pub current_turn: String,
+    pub button_player: String,
+    pub min_bet: u128,
 }
